Refactor API route to use execId as a string instead of ObjectId
This commit modifies the PUT function in the API route for execs. It changes the query parameter _id from being an ObjectId to a string. This change simplifies the logic and improves compatibility with the database. Additionally, it adds a config object to enable correct parsing of the request body.

Refactor mongodb.ts to manage MongoDB connection in development

This commit updates the mongodb.ts file to manage the MongoDB connection differently based on the environment. In development, it uses a global variable to avoid opening multiple connections. This change improves efficiency and prevents unnecessary resource usage.